pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gherkin-verifier"
version = "0.1.0"
description = "Formal verifier for logical contradictions in Gherkin specifications"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "z3-solver>=4.12.0",
    "gherkin-official>=28.0.0",
    "networkx>=3.0",
    "sympy>=1.12",
]

[project.scripts]
gherkin-verify = "gherkin_verifier.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gherkin_verifier*"]
